pyproject.toml
[project]
name = "cooloc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiopg>=1.4.0",
    "fastapi>=0.115.12",
    "logger>=1.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-extra-types>=2.10.3",
    "pydantic-settings>=2.9.1",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "python-ulid>=3.0.0",
    "uvicorn>=0.34.2",
]
